proof-of-concept/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
PyYAML==6.0
tomli==2.0.1
